3 byte fields (#32)
* Update fromfile to support 3-byte fields

Some FCS files (such as the one generated by the Cytek xP5) store integers in 3-byte fields.  This breaks numpy's parser, which only wants power-of-two sized fields.  So, I've updated fromfile() to parse FCS files as a table of 1-byte unsigned ints, expand those fields to 4 bytes, and then re-view them as the proper dtype.

My only concern is that this may break on an FCS file produced by a big-endian machine (because I'm adding the extra bytes to the beginning of the field).  I don't have ready access to any of those -- do you?

* add tests for 3-byte DATA

* documentation for fromfile() changes