Make __init__.py use a relative import (#42)
Instead of making `cytoflow` depend on `fcsparser`, I would like to include `fcsparser` as a submodule. This way, if I have fixes that I want to roll out with a `cytoflow` release, I don't have to wait on you to roll a new release of `fcsparser`.  (Making new releases is harder than it should be.)

This would be a lot easier on my end if `fcsparser.__init__` used a relative import instead of an absolute import. This way, the directory containing `api.py` doesn't have to be in `sys.path`.  It should not effect folks who use `fcsparser` directly.